MailBoxType CoroutineLockType LocationType 都改成int，都带上PackageType，避免Package之间冲突

+++ Unity/Packages/cn.etetet.actorlocation/Scripts/Hotfix/Server/MailBoxType_OrderedMessageHandler.cs
@@ -1,6 +1,6 @@
 ﻿namespace ET
 {
-    [Invoke((long)MailBoxType.OrderedMessage)]
+    [Invoke(MailBoxType.OrderedMessage)]
     public class MailBoxType_OrderedMessageHandler: AInvokeHandler<MailBoxInvoker>
     {
         public override void Handle(MailBoxInvoker args)
