+++ Unity/Packages/cn.etetet.core/Scripts/Hotfix/Share/MailBoxType_UnOrderedMessageHandler.cs
@@ -1,6 +1,6 @@
 ﻿namespace ET
 {
-    [Invoke((long)MailBoxType.UnOrderedMessage)]
+    [Invoke(MailBoxType.UnOrderedMessage)]
     public class MailBoxType_UnOrderedMessageHandler: AInvokeHandler<MailBoxInvoker>
     {
         public override void Handle(MailBoxInvoker args)
