+++ Unity/Packages/cn.etetet.statesync/Scripts/Hotfix/Server/Gate/MailBoxType_GateSessionHandler.cs
@@ -1,6 +1,6 @@
 ﻿namespace ET.Server
 {
-    [Invoke((long)MailBoxType.GateSession)]
+    [Invoke(MailBoxType.GateSession)]
     public class MailBoxType_GateSessionHandler: AInvokeHandler<MailBoxInvoker>
     {
         public override void Handle(MailBoxInvoker args)
